Move trait impls that simply forward to something else to their own module
